pyproject.toml
[project]
name = "polars-polars-polars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "boto3>=1.42.2",
    "google-cloud-secret-manager>=2.25.0",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-exporter-otlp-proto-http>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    "polars>=1.35.2",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-xdist>=3.8.0",
    "rich>=14.2.0",
    "ruff>=0.14.7",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.7",
]

[tool.pytest.ini_options]
testpaths = [ "src/tests" ]
pythonpath = [ "." ]

[tool.ruff]
lint.select = ["I"]